src/insights_sender/requirements.txt
boto3==1.34.154
botocore==1.34.154
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redmail==0.6.0
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
